+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Option to change random size.
+
+### Changed
+- Change default random size to 32 bytes (256 bits).
 
 ## [0.1.1] - 2018-11-03
 ### Added